Simplifying Workflow: Auto-Process Paragraph at Cursor
The prerequisite for executing commands in this plugin is that the user has selected some text in the active document. 

However, common practice suggests that the entire paragraph should be processed. In such cases, adding an operation to select text becomes very cumbersome and inconvenient. 

I have modified the main.js file to primarily achieve the following functionality: when the user group executes the command, if no text has been selected in advance, it is assumed that the user intends to process the paragraph where the cursor is located. 

The system will help the user select the entire paragraph by default. 

I hope this can be accepted. Wishing you all the best.